Feat: 카카오 신규회원 비번변경 API 추가

+++ src/main/java/com/starta/project/domain/member/service/MemberService.java
@@ -100,6 +100,19 @@ public class MemberService {
 
     }
     @Transactional
+    public MsgResponse kakaoFirstLogin(KaKaoFirstLoginDto requestDto, Long id) {
+        Member member = validationUtil.findMember(id);
+        String newPassword = requestDto.getNewPassword();
+        validationUtil.checkPassword(newPassword, requestDto.getNewCheckPassword());
+        if (!passwordEncoder.matches(requestDto.getPassword(), member.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+        String encodedPassword = passwordEncoder.encode(newPassword);
+        member.updatePassword(encodedPassword);
+        return new MsgResponse("카카오 신규유저 비밀번호 변경완료");
+    }
+
+    @Transactional
     public MsgResponse updatePassword(UpdatePasswordRequestDto requestDto, Long id) {
         Member member = validationUtil.findMember(id);
         validationUtil.checkPassword(requestDto.getNewPassword(), requestDto.getNewCheckPassword());
@@ -123,10 +136,12 @@ public class MemberService {
 
     @Transactional  // 일관성 유지를 위해 사용
     public MsgResponse deleteMember(String password, Member member) {
-//        if (!passwordEncoder.matches(password, member.getPassword())) {
-//            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
-//        }
+        if (!passwordEncoder.matches(password, member.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
         memberRepository.delete(member);
         return new MsgResponse("탈퇴완료.");
     }
+
+
 }
